feat: Extract status to typing module
