+++ falcon/errors.py
@@ -35,17 +35,15 @@ package namespace::
 """
 
 from datetime import datetime
-from typing import Dict
 from typing import Iterable
 from typing import List
 from typing import Optional
-from typing import Tuple
 from typing import Union
 
 from falcon.http_error import HTTPError
 import falcon.status_codes as status
-from falcon.typing import NormalizedHeaders
-from falcon.typing import RawHeaders
+from falcon.typing_http_data import NormalizedHeaders
+from falcon.typing_http_data import RawHeaders
 from falcon.util.deprecation import deprecated_args
 from falcon.util.misc import dt_to_http
