+++ falcon/typing.py
@@ -13,11 +13,7 @@
 # limitations under the License.
 
 """Shorthand definitions for more complex types."""
-
-from typing import Any, Callable, Pattern, Union
-from typing import Dict
-from typing import List
-from typing import Tuple
+from typing import Any, Callable, Pattern
 from typing import Union
 
 from falcon.request import Request
@@ -38,5 +34,3 @@ SinkPrefix = Union[str, Pattern]
 #   arguments afterwords?
 # class SinkCallable(Protocol):
 #     def __call__(sef, req: Request, resp: Response, <how to do?>): ...
-NormalizedHeaders = Dict[str, str]
-RawHeaders = Union[NormalizedHeaders, List[Tuple[str, str]], None]
