+++ falcon/typing_http_data.py
@@ -1,0 +1,9 @@
+import http
+from typing import Dict
+from typing import List
+from typing import Tuple
+from typing import Union
+
+NormalizedHeaders = Dict[str, str]
+RawHeaders = Union[NormalizedHeaders, List[Tuple[str, str]], None]
+Status = Union[http.HTTPStatus, str, int]
